Added $restBases property.
Why? To avoid expensive process of checking empty and running array_keys
when getPostTypeRestBases() more than once.

Think about it.  The post types are stored away when the object is created.
Therefore, it makes sense to extract the reset_base from each and cache it
for use within the object.

Results:

1. Runs faster.
2. But increases memory usage as the object now holds the array.