Changed `shouldFilter()` into a guard clause.
The intent of the method `shouldFilter()` is to protect the
callback by determine if there is a "no go" condition.  If
one is found, then the callback should bail out.

Changes in this commit are:

1. Moves `is_null()` check to `shouldFilter()`.
2. Injects the given input from posts_pre_query filter.
3. Creates a new private method called `doingREST()`.  The
intent of this method is to determine if WordPress is doing a
RESTful request or not.
4. Removes the inline comment as it is no longer needed. The
method's name clearly identifies its intent.

@see Code Review Part 1 article on torque
@link https://torquemag.io/2018/04/code-review-part-1-fixing-design-flaw-return-early-strategy/