Abstracts post generator.
The task of generating posts is a separate feature. Therefore,
it deserves to have its own class.

Though this plugin is for educational purposes, the task of
generating posts is part of a normal production model.  Here we
are generating mocked posts.  But in a real plugin, you would likely
fetch from the database or generate a series of new posts based
upon certain criteria.

By abstracting the "generation" process, we focus the FilterWPQuery
class on its job: to filter WP_Query.