Defined the embedded variable.
PHP needs your help to quickly identify the variable that is
embedded inside of a string.  Wrapping the variable inside of
curly braces tells the PHP parser: "Hey, this is an embedded
variable."

Why?

Imagine that the title had another character after the variable.
How would PHP know that the variable is `$postNumber` and not that
plus the other character(s).

The parser is greedy.  And a string is comprised of different
characters.  Where does a variable start and end?

The better practice of wrapping an embedded variable in curly
braces is expressive and declarative.

As an added benefit, it helps you and me to quickly identify
it as variable to understand that its value will be swapped out
and placed in the post's new title.

@see http://php.net/manual/en/language.types.string.php#language.types.string.parsing