Removed unnecessary assignment.
The variable is not used.  It's not being evaluated or further
processed.  Therefore, it is not necessary.

In addition, this change makes the code more performant. Why?
It requires less steps and less memory.

Before, here's the sequence:

1. Control flows to getPosts().
2. Control flows back to filterPreQuery().
3. The returned array is assigned to a variable.
- The variable is added to the PHP symbols table.
- The array is added into the function's scope, creating a new memory
bucket for it.
- The variable and array are bound together, meaning that the
variable points to the memory location of the array.
4. The control flows to the next line of code.
5. PHP looks up the variable in its symbols table.
6. It exits the method and returns the looked up array to the caller.

After the change, we eliminate:

1. the variable assignment
2. the mapping between the variable and the array's memory location.
3. the extra control flow step, i.e. #4 above.
4. the variable lookup.

The code has less steps to process, making it more performant.

The code is also more readable as one does not have to wonder why
the variable exists and how it might be used.  Less code is good.