Replaced hard-coded priority with the property.
The class defines the priority number.  It's redundant to
hard-code the priority number in the `add_filter` and `remove_filter`
tasks.  This change makes the code DRY.

What are the advantages here?

The priority number can change in one place, i.e. in the property.
This is important as you may want to change the priority.  But you
might forget to change it in each of the instances. Whoopsie.  That
will create a bug.

Using the new strategy, the priority number is set in only one place.
Then the code gets that value when doing its work.