Removed scrutinizer's ignore annotation.
Scrutinizer allows you to ignore certain checks with an inline
annotation.  But this annotation has to come immediately before
the WordPress function.

The problem here is this:

It makes the code less readable.

That is not what we want.

By removing the annotations, we improve the code's readability, but
Scrutinizer fails, marking each as a major bug.

How do we fix Scrutinizer?

There are 2 options:

1. Teach Scrutinizer via it's UI. Yuck.
2. Include WordPress as a composer dependency.

@see https://scrutinizer-ci.com/docs/tools/php/php-scrutinizer/analyzing-wordpress