Move the Null Checker.
Our goal is to convert the shouldFilter() method into a
decision maker to determine if the code should filter or not.

Step 1 is to move the null checker into this method and pass the
input to it.

By doing this step, we can remove the inline comments and rely
on the shouldFilter() method to be the 'return early' decision
maker.