Abstracted the REST checker to improve readability.
Step 3 is create a new private method for the REST checker.

Why?

Without the inline comment, the conditional's intent is not clear.
did_action('rest_api_init') does not tell us what this checker is
doing.  It just tells us that the 'rest_api_init' event fired.

Our goal is to build our code to be highly human readable, such
that anyone (including you) can quickly understand what's going on.
We design code to _tell us_ what's happening and why.

By moving this function to a new method, we use the name of the
method to tell us it's intent, i.e. it's checking if we are
doing a REST request.