Skip the assignment and just return.
The assignment variable is unnecessary, as we are not using the
variable other than to return the value it represents.  Making
an unnecessary assignment makes the code less performant.  Why?

1. PHP has to create the new variable in its symbols table.
2. It has to bind that spot to the value's memory location.
3. Then it shifts control to the next line.
4. Next, it has to lookup the value for that variable before
returning.

These extra steps add memory and increase the code's execution
time.

By refactoring to just return, we eliminate these extra steps.
The code will run faster.  But it is also more readable, making it
more maintainable.