Renamed "callback" to tell us what its intent.
The word "callback" is generic.  It does tell us that it is
bound to something that will invoke it.  But it does not
tell us what it will do, i.e. what its behavior is when it's
called.

This "callback" job is to get the posts for a RESTful request.
It's bound to a filter, i.e. `post_pre_query`, which is fired
_before_ the query takes place.

When figuring out a name, we first identify what it does:

1. It's a filter, which changes the input for 'post_pre_query' event.
2. It gets or generates an array of posts to be used for the query.

Using a name such as `filterPreQuery` tells us that when this
method is called, it will "filter the pre query.