Abstract posts generation to generator method.
The getPosts() method's intent to fetch the posts for the
pre-query. In a real world plugin, it's job is to be a
handler or manager.  It decides which mechanism to use to
either get or generate posts.

That means the job of dynamically generating posts is a
different task.

The inline comments are there as a neon light telling us
to refactor.  By abstracting the post generator to its
own method, we can:

1. get rid of the inline comments
2. set up the class to be a handler or manager, meaning that
we can extend the functionality to load different implementations
based upon our business rules and given inputs.