Removed unnecessary parentheses from post instantiation.
The keyword `new` creates an object for us.  The only time we
need to wrap an object inside of a pair of parentheses is when
we want to chain a method directly to the object without assigning
the object to a variable.

Here, removing the unnecessary () makes the code more readable.

Remember, formatting matters. :)