Removed the unnecessary array index.
The posts start at index 0 and so does an indexed array.
Therefore, there is no need to specifically specify the new
post's position in the array.

This change simplifies the code and makes it more readable.
But it also makes it slightly more performant as PHP does not
have to lookup the value presented by variable.