Added an init() method & injected the implementation.
Notice that we are type hinting the injected object using the
contract.  This contract serves an interface between the FilterWPQuery
and each implementation.

Using this design, the FilterWPQuery class does not care which
implementation is wired to it.  Rather, it serves as a generic object.
But because of the contract, the FilterWPQuery knows how to interface
with it.